requirements.txt
minimalmodbus
psycopg[c]
py-vigil-reporter
